+++ src/data/__init__.py
@@ -1,5 +1,7 @@
 from .idatabase import IDatabase
+from .sqlitedatabase import SQLiteDatabase
 
 __all__ = [
     IDatabase,
+    SQLiteDatabase
 ]